refactor servlets for more clarity
